+++ src/Core/ListType.ts
@@ -1,0 +1,14 @@
+/**
+ * Represents a list-type.
+ */
+export enum ListType
+{
+    /**
+     * An unordered list.
+     */
+    ul,
+    /**
+     * An ordered list.
+     */
+    ol
+}